backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.115.11",
    "google-generativeai>=0.8.3",
    "langchain-anthropic>=0.3.10",
    "langchain-community>=0.3.20",
    "langchain-google-genai>=2.1.1",
    "langchain-mistralai>=0.2.9",
    "langchain-neo4j>=0.4.0",
    "langchain-openai>=0.3.10",
    "langgraph>=0.3.18",
    "pydantic>=2.10.6",
    "python-dotenv>=1.0.1",
    "pypdf>=4.0.0",
    "pdfplumber>=0.10.0",
    "python-multipart>=0.0.6",
    "redis>=5.0.0",
    "openinference-instrumentation-langchain>=0.1.0",
    "opentelemetry-sdk>=1.20.0",
    "opentelemetry-exporter-otlp>=1.20.0",
    "langchain>=0.3.0",
    "langchain-text-splitters>=0.3.0",
    "mcp>=0.1.0",
    "fastmcp>=0.1.0",
    "nest_asyncio>=1.6.0",
]
